app/requirements.txt
streamlit
requests
pillow
